+++ packages/agent-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @clawmasons/agent-sdk
 
+## 0.1.7
+
+### Patch Changes
+
+- Patch bump all packages
+- Updated dependencies
+  - @clawmasons/shared@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes